Don't override background-images for all inlines in forced colors
In a previous change, we had ignored background images for all inlines
to prevent background images from being painted on top of the
readability backplate in forced colors mode.
https://chromium-review.googlesource.com/c/chromium/src/+/1757685

However, it turns out that inline-block, inline-table, inline-flex,
and inline-grid paint their backgrounds before the backplate is
painted. As such, we do not want to ignore background images for these
inlines in forced colors mode.

Bug: 970285
Change-Id: Iee2463775e67a1661cbb85f0a51b5bb8f01566cb