Use the portal-host-including parent for console logging in AncestorThrottle.
This is already used for determining whether ancestors are permitted.
It needs to be used also for figuring out which ancestor's console to
log in.

A web test that reaches this code using X-Frame-Options: DENY is included.
Without the change to content/, this test fails with the same stack trace
as in the linked bug.

Bug: 1038256
Change-Id: Iad3623d82e0b2192eab626a2184d09d9d3e7be62